fix : Board,RacingInfo class dto 패키지로 refactor
